redundant files deleted
