+++ app/src/main/AndroidManifest.xml
@@ -36,19 +36,18 @@
         android:hardwareAccelerated="false"
         android:largeHeap="true"
         android:theme="@style/AppTheme.NoActionBar">
-        <activity android:name=".GameInfoActivity" />
         <meta-data
             android:name="com.google.android.actions"
             android:resource="@xml/network_security_config" />
         <activity android:name=".GameActivity" />
         <activity android:name=".ProfileActivity" />
         <activity android:name=".Gallery.GalleryProvider" />
-        <activity android:name=".ForgotPasswordActivity" />
-        <activity android:name=".RegisterHandler"/>
-        <activity android:name=".ForgotPasswordHandler" />
-        <activity android:name=".LoginHandler" />
-        <activity android:name=".ForgotPasswordTokenActivity" />
-        <activity android:name=".ForgotPasswordTokenHandler" />
+        <activity android:name=".Entrance.ForgotPasswordActivity" />
+        <activity android:name=".EntranceHandlers.RegisterHandler"/>
+        <activity android:name=".EntranceHandlers.ForgotPasswordHandler" />
+        <activity android:name=".EntranceHandlers.LoginHandler" />
+        <activity android:name=".Entrance.ForgotPasswordTokenActivity" />
+        <activity android:name=".EntranceHandlers.ForgotPasswordTokenHandler" />
         <activity android:name=".User"/>
         <activity android:name=".URLs" />
         <activity
@@ -61,8 +60,8 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/map_key" />
 
-        <activity android:name=".RegisterActivity" />
-        <activity android:name=".LoginActivity" />
+        <activity android:name=".Entrance.RegisterActivity" />
+        <activity android:name=".Entrance.LoginActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
