+++ app/src/main/java/com/example/citygame/Entrance/ForgotPasswordTokenActivity.java
@@ -1,4 +1,4 @@
-package com.example.citygame;
+package com.example.citygame.Entrance;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,7 +9,10 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import java.util.Arrays;
+import com.example.citygame.EntranceHandlers.ForgotPasswordTokenHandler;
+import com.example.citygame.MainActivity;
+import com.example.citygame.R;
+import com.example.citygame.User;
 
 
 public class ForgotPasswordTokenActivity extends AppCompatActivity {
