+++ app/src/main/java/com/example/citygame/EntranceHandlers/ForgotPasswordTokenHandler.java
@@ -1,22 +1,27 @@
-package com.example.citygame;
+package com.example.citygame.EntranceHandlers;
 
 import android.os.AsyncTask;
 
+import com.example.citygame.URLs;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
 import java.io.BufferedReader;
+import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
 
-public class LoginHandler extends AsyncTask<String, Void, Boolean> {
+public class ForgotPasswordTokenHandler extends AsyncTask<String, Void, Boolean> {
 
-    private URLs urlGet = new URLs();
-    private LoginHandlerFinishedListener loginListener;
+    private URLs urlPost = new URLs();
+    private ForgotPasswordTokenHandler.ForgotPasswordTokenHandlerFinishedListener listener;
 
-
-    public LoginHandler(LoginHandlerFinishedListener loginListener) {
-        this.loginListener = loginListener;
+    public ForgotPasswordTokenHandler(ForgotPasswordTokenHandler.ForgotPasswordTokenHandlerFinishedListener listener) {
+        this.listener = listener;
     }
 
     @Override
@@ -24,16 +29,27 @@ public class LoginHandler extends AsyncTask<String, Void, Boolean> {
         HttpURLConnection connection = null;
         StringBuilder stringBuilder = new StringBuilder();
         try {
-            String urlAppended = new StringBuilder(urlGet.getServerURLLogin()).append("?password=").append(strings[0]).append("&email=").append(strings[1]).toString();
-            URL url = new URL(urlAppended);
+            URL url = new URL(urlPost.getServerURLPasswordReset());
+            DataOutputStream outputStream;
             connection = (HttpURLConnection) url.openConnection();
             connection.setDoInput(true);
-            connection.setRequestMethod("GET");
+            connection.setDoOutput(true);
+            connection.setRequestMethod("POST");
             connection.setUseCaches(false);
             connection.setConnectTimeout(10000);
             connection.setReadTimeout(10000);
             connection.setRequestProperty("Content-Type", "application/json");
+            connection.setRequestProperty("Accept", "application/json");
             connection.connect();
+
+            JSONObject jsonParam = new JSONObject();
+            jsonParam.put("token", strings[0]);
+            jsonParam.put("password", strings[1]);
+
+            outputStream = new DataOutputStream(connection.getOutputStream());
+            outputStream.writeBytes(jsonParam.toString());
+            outputStream.flush();
+            outputStream.close();
 
             int httpResult = connection.getResponseCode();
 
@@ -58,18 +74,29 @@ public class LoginHandler extends AsyncTask<String, Void, Boolean> {
             e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
+        } catch (JSONException e) {
+            e.printStackTrace();
         }
         return false;
     }
 
     @Override
-    protected void onPostExecute(Boolean result) {
-       if (this.loginListener == null) return;
-       this.loginListener.onFinished(result);
+    protected void onPreExecute() {
+        super.onPreExecute();
     }
 
-    public interface LoginHandlerFinishedListener {
+    @Override
+    protected void onProgressUpdate(Void... values) {
+        //findViewById(R.id.loadingPanel).setVisibility(View.GONE);
+    }
+
+    @Override
+    protected void onPostExecute(Boolean result) {
+        if (this.listener == null) return;
+        this.listener.onFinished(result);
+    }
+
+    public interface ForgotPasswordTokenHandlerFinishedListener {
         void onFinished(Boolean result);
     }
-
 }
