+++ app/src/main/java/com/example/citygame/EntranceHandlers/LoginHandler.java
@@ -1,57 +1,41 @@
-package com.example.citygame;
+package com.example.citygame.EntranceHandlers;
+
 import android.os.AsyncTask;
 
-import org.json.JSONException;
-import org.json.JSONObject;
+import com.example.citygame.URLs;
 
 import java.io.BufferedReader;
-import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
 
+public class LoginHandler extends AsyncTask<String, Void, Boolean> {
 
-public class RegisterHandler extends AsyncTask<String, Void, Boolean> {
+    private URLs urlGet = new URLs();
+    private LoginHandlerFinishedListener loginListener;
 
-    private URLs urlPost = new URLs();
-    private RegistrationHandlerFinishedListener listener;
 
-    public RegisterHandler(RegistrationHandlerFinishedListener listener) {
-        this.listener = listener;
+    public LoginHandler(LoginHandlerFinishedListener loginListener) {
+        this.loginListener = loginListener;
     }
 
     @Override
     protected Boolean doInBackground(String... strings) {
-
         HttpURLConnection connection = null;
         StringBuilder stringBuilder = new StringBuilder();
         try {
-            URL url = new URL(urlPost.getServerURLRegistration());
-            DataOutputStream outputStream;
+            String urlAppended = new StringBuilder(urlGet.getServerURLLogin()).append("?password=").append(strings[0]).append("&email=").append(strings[1]).toString();
+            URL url = new URL(urlAppended);
             connection = (HttpURLConnection) url.openConnection();
             connection.setDoInput(true);
-            connection.setDoOutput(true);
-            connection.setRequestMethod("POST");
+            connection.setRequestMethod("GET");
             connection.setUseCaches(false);
             connection.setConnectTimeout(10000);
             connection.setReadTimeout(10000);
             connection.setRequestProperty("Content-Type", "application/json");
-            connection.setRequestProperty("Accept", "application/json");
             connection.connect();
-
-            // create JSONObject here
-            JSONObject jsonParam = new JSONObject();
-            jsonParam.put("password", strings[0]);
-            jsonParam.put("email", strings[1]);
-            jsonParam.put("name", strings[2]);
-
-            // send POST output
-            outputStream = new DataOutputStream(connection.getOutputStream());
-            outputStream.writeBytes(jsonParam.toString());
-            outputStream.flush();
-            outputStream.close();
 
             int httpResult = connection.getResponseCode();
 
@@ -76,32 +60,18 @@ public class RegisterHandler extends AsyncTask<String, Void, Boolean> {
             e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
-        } catch (JSONException e) {
-            e.printStackTrace();
         }
         return false;
     }
 
     @Override
-    protected void onPreExecute() {
-        super.onPreExecute();
-    }
-
-    @Override
-    protected void onProgressUpdate(Void... values) {
-        //findViewById(R.id.loadingPanel).setVisibility(View.GONE);
-    }
-
-    @Override
     protected void onPostExecute(Boolean result) {
-        if (this.listener == null) return;
-        this.listener.onFinished(result);
+       if (this.loginListener == null) return;
+       this.loginListener.onFinished(result);
     }
 
-    public interface RegistrationHandlerFinishedListener {
+    public interface LoginHandlerFinishedListener {
         void onFinished(Boolean result);
-
     }
-
 
 }
