+++ app/src/main/java/com/example/citygame/MainActivity.java
@@ -11,11 +11,12 @@ import android.widget.Button;
 import android.widget.PopupMenu;
 import android.widget.Toast;
 
+import com.example.citygame.Entrance.LoginActivity;
+import com.example.citygame.Entrance.RegisterActivity;
 import com.example.citygame.MarkersList.Marker;
 import com.example.citygame.MarkersList.MarkerListActivity;
 
 import java.util.Arrays;
-import java.util.List;
 
 public class MainActivity extends AppCompatActivity implements PopupMenu.OnMenuItemClickListener {
