+++ app/src/main/res/layout/activity_forgot_password_token.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".ForgotPasswordTokenActivity">
+    tools:context=".Entrance.ForgotPasswordTokenActivity">
 
     <EditText
         android:id="@+id/PasswordOneEditText"
