+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@ xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 android:background="@drawable/bg"
-tools:context=".LoginActivity">
+tools:context=".Entrance.LoginActivity">
 
 <EditText
     android:id="@+id/LoginEditText"
